ChatGPT told me to refactor the whole world
